+++ .gitignore
@@ -1,1 +1,5 @@
+#ignoring node_modules
 /node_modules
+
+#ignoring build module
+/public/bundle.js
