+++ public/entry.js
@@ -1,0 +1,1 @@
+import inf_scroller from './inf-scroller';
