+++ public/index.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+    <link rel="stylesheet" href="style.css">
+</head>
+<body>
+    <div class="scroller">
+        <ul class="tag-list scroller-inner">
+            <!--insert elem-->>
+        </ul>
+        <ul class="tag-list scroller-inner">
+            <!--inert elem-->>
+        </ul>
+    </div>
+    <script type="module" src="bundle.js"></script>
+</body>
+</html>
